Bit of cleanup, moved main loop for Virtual Machine into its own file C#
